Fix public/ namespace in tests
